Add files via upload
Mastering Banking offers a succinct account of the basic features of banking, and is designed for the reader who is new to the subject. The second edition has been fully revised and up-dated to cover the new world of retail banking in the UK financial markets, and to meet the requirements of all current courses at this level. Ideal for CIB Banking Certificate Part 1, CIB Pre-Associateship Route, NVQ Levels 2 and 3, and for students on business and secretarial courses at GNVQ, GCSE and A-Level. Associate banking students and the established banker will also find this book an invaluable refresher.